normalize rectangle object
