+++ src/burn/drv/pst90s/d_ssv.cpp
@@ -9,6 +9,7 @@
 #include "eeprom.h"
 #include "math.h"
 #include "burn_gun.h"
+#include "rectangle.h"
 
 // Use for slower systems like rpi & xbox
 #define SSV_UPD_SPEEDHACK
@@ -17,33 +18,6 @@
 	srmp7 - no music/sfx
 	analog inputs not hooked up at all
 */
-
-struct rectangle
-{
-	INT32 min_x;
-	INT32 max_x;
-	INT32 min_y;
-	INT32 max_y;
-
-	rectangle(INT32 minx = 0, INT32 maxx = 0, INT32 miny = 0, INT32 maxy = 0) {
-		set(minx, maxx, miny, maxy);
-	}
-
-	void set(INT32 minx, INT32 maxx, INT32 miny, INT32 maxy) {
-		min_x = minx; max_x = maxx;
-		min_y = miny; max_y = maxy;
-	}
-
-	rectangle operator &= (const rectangle &other) {
-		if (min_x < other.min_x) min_x = other.min_x;
-		if (min_y < other.min_y) min_y = other.min_y;
-		if (max_x > other.max_x) max_x = other.max_x;
-		if (max_y > other.max_y) max_y = other.max_y;
-		if (min_y > max_y) min_y = max_y;
-		if (min_x > max_x) min_x = max_x;
-		return *this;
-	}
-};
 
 static UINT8 *AllMem;
 static UINT8 *MemEnd;
